fix(rename): ディレクトリ名(atoms -> ui)と型名(Api -> HoloData)を変更
