+++ pages/_app.tsx
@@ -3,9 +3,9 @@ import type { AppProps } from "next/app"
 import { useRouter } from "next/router"
 import { useEffect } from "react"
 import * as gtag from "../lib/gtag"
-import { GaScript } from "../components/atoms/gaScript"
+import { GaScript } from "../components/ui/gaScript"
 import ChangeCardSizeObserver from "../utils/globalChangeCardObserver"
-import HeadMeta from "../components/atoms/headMeta"
+import HeadMeta from "../components/ui/headMeta"
 
 function MyApp({ Component, pageProps }: AppProps) {
   const router = useRouter()
